Added Arduino Long lost Arduino Code
Finally added the Arduino side of the UT Running code. It includes 3
Arduino applications:
1] BluetoothConfig: Used to configure various settings of the Bluetooth
module. NOTE: This code connects to the Bluetooth Module directly over
UART to configure it. As this is the case, this code cannot not be used
properly on the LilyPad, only on an actual Arduino that has multiple
UART's, one connected to the USB port and another to the Bluetooth
Module itself.

2] BluetoothTestLilyPad: Doesn't do much, but is a good example of how
to program the Bluetooth Module

3] UntetheredBT: This is the actual code that is running on the
LilyPad. It handles intercepting Bluetooth packets and controlling the
buzzers accordingly.